Create README.md
A simple and useful online store
This project is a web-based product management application. The site includes functionality for adding and removing products, as well as a convenient filter for quickly searching for products by name.

Ease of use: The interface is intuitive, allowing even beginners to easily work with the application.